Sophgo: Fix dmidecode display errors for cache and DDR size
- Type07: Modify the logic for filling Type 7 InstalledSize
  and InstalledSize2 according to the SMBIOS 3.1 specification
  and dmidecode source code to ensure that dmidecode correctly
  parses the cache size.

- Type17: Modify the logic for filling Type 17 Size and ExtendedSize
  according to the SMBIOS 3.1 specification. Additionally, update
  the menu display logic for DDR size conversion to align with dmidecode.

+++ Silicon/Sophgo/Include/Library/SmbiosInformationLib.h
@@ -35,9 +35,9 @@ typedef struct {
   UINT16 ProcessorCurrentSpeed;
   UINT8  ProcessorCoreCount;
   UINT8  ProcessorThreadCount;
-  UINT16 L1ICacheSize;
-  UINT16 L1DCacheSize;
-  UINT16 L2CacheSize;
+  UINT32 L1ICacheSize;
+  UINT32 L1DCacheSize;
+  UINT32 L2CacheSize;
   UINT32 L3CacheSize;
   UINT8 InstallableLanguages;
   UINT8 BiosLanguageFlags;
